web: Common: Renamed folder advanced-selecto2 to AdvancedSelector
