+++ web/ASC.Web.Common/src/components/AdvancedSelector/AdvancedSelector.js
