+++ web/ASC.Web.Common/src/components/AdvancedSelector/index.js
