+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/Footer.js
