+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/Header.js
