+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/Selector.js
