+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/StyledBody.js
