+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/StyledColumn.js
