+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/StyledHeader.js
