+++ web/ASC.Web.Common/src/components/AdvancedSelector/sub-components/StyledSelector.js
