+++ web/ASC.Web.Common/src/components/index.js
@@ -6,4 +6,4 @@ export { default as ModuleTile } from "./module-tile";
 export { default as Headline } from "./headline";
 export { default as PeopleSelector } from "./people-selector";
 export { default as GroupSelector } from "./group-selector";
-export { default as AdvancedSelector2 } from './advanced-selector2'
+export { default as AdvancedSelector } from './AdvancedSelector'
